+++ CHANGELOG.md
@@ -26,6 +26,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 ### Fixed
 
 - Apply custom error handlers for syntax or request errors https://github.com/nuwave/lighthouse/pull/1726
+- Define scalars instead of `Mixed` type in directive definitions https://github.com/nuwave/lighthouse/pull/1742
 
 ## 5.2.0